Add files via upload
A/B tests are very commonly performed by data analysts and data scientists. It is important that you get some practice working with the difficulties of these

For this project, I analyzed the results of an A/B test run by an e-commerce website. To help the company understand if they should implement the new page, keep the old page, or perhaps run the experiment longer to make their decision.